+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Fix order id lookup
+### Changed
+- Functions to support Payer Auth
+
 ## [1.6.2] - 2022-10-06
 
 ### Changed